+++ src/pages/app/skills.tsx
@@ -1,10 +1,13 @@
-import SkillsForm from '../../components/SkillsForm'
+import SkillSelect from '../../components/SkillSelect'
+import { useState } from 'react'
 import fetch from 'node-fetch'
 
 export default function Skills() {
+  const [skills, setSkills] = useState(null)
   return (
     <>
-      <SkillsForm></SkillsForm>
+      <SkillSelect setSkills={setSkills}></SkillSelect>
+      {console.log(skills)}
     </>
   )
 }
